+++ lib/arel/extensions.rb
@@ -51,6 +51,7 @@ require 'arel/extensions/default_values'
 require 'arel/extensions/conflict'
 require 'arel/extensions/infer'
 require 'arel/extensions/set_to_default'
+require 'arel/extensions/update_statement'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
