+++ lib/arel/extensions/nodes.rb
@@ -5,13 +5,6 @@
 module Arel
   module Nodes
     # https://www.postgresql.org/docs/10/sql-update.html
-    Arel::Nodes::UpdateStatement.class_eval do
-      attr_accessor :with
-      attr_accessor :froms
-      attr_accessor :returning
-    end
-
-    # https://www.postgresql.org/docs/10/sql-update.html
     class CurrentOfExpression < Arel::Nodes::Node
       attr_accessor :cursor_name
